Show warning when attaching a debugger to a non-debuggable APK.
Currently a user can use the "Debug" action to attach debuggers to
any APK built by an android_binary target and debugger attachment will
fail if the APK is not debuggable (unless they are deploying to a
userdebug device).  The IDE now detects this and warns the user by
instructing them to set "android:debuggable" to true in the manifest.

PiperOrigin-RevId: 292662798